+++ src/constants.js
@@ -1,4 +1,3 @@
-export const FB_API_VERSION = 'v2.9';
 export const BUTTON_TYPES = [
   'web_url',
   'postback',